tg_client.py
"""
Telegram client module using Telethon library.

Provides functionality to connect to Telegram, list channels,
select a channel, and download messages for processing.

TODO: Implement Telethon client connection
TODO: Add session management for persistent login
TODO: Add channel listing functionality
TODO: Add interactive channel selection
TODO: Add message downloading with pagination
TODO: Add JSON export functionality
TODO: Add error handling for network issues
TODO: Add rate limiting to avoid API restrictions
TODO: Add progress reporting for large downloads
TODO: Add message filtering options
"""

import os
from typing import List, Dict, Optional
from pathlib import Path
from telethon import TelegramClient
from telethon.tl.types import Channel, Chat
from dotenv import load_dotenv

# Load environment variables
load_dotenv()


def get_telegram_credentials() -> tuple[str, str]:
    """
    Get Telegram API credentials from environment variables.
    
    Returns:
        Tuple of (api_id, api_hash)
        
    Raises:
        ValueError: If credentials are not found
    """
    api_id = os.getenv('TELEGRAM_API_ID')
    api_hash = os.getenv('TELEGRAM_API_HASH')
    
    if not api_id or not api_hash:
        raise ValueError(
            "Telegram credentials not found. Please set TELEGRAM_API_ID and "
            "TELEGRAM_API_HASH in your .env file or environment variables."
        )
    
    return api_id, api_hash


def get_session_name() -> str:
    """
    Get Telegram session name from environment or use default.
    
    Returns:
        Session name for Telegram client
    """
    return os.getenv('TELEGRAM_SESSION_NAME', 'tg2kb_session')


def connect_telethon(api_id: str, api_hash: str) -> TelegramClient:
    """
    Establish connection to Telegram using Telethon.
    
    Args:
        api_id: Telegram API ID from https://my.telegram.org
        api_hash: Telegram API Hash from https://my.telegram.org
        
    Returns:
        Connected TelegramClient instance
        
    TODO: Initialize TelegramClient with session file
    TODO: Handle connection errors
    TODO: Implement session persistence
    TODO: Add connection timeout handling
    TODO: Add retry logic for failed connections
    """
    # TODO: Create TelegramClient instance
    # TODO: Connect to Telegram
    # TODO: Handle authentication (phone number, code)
    # TODO: Save session for future use
    return None  # Placeholder return


def get_user_channels(client: TelegramClient) -> List[Dict]:
    """
    Get list of user's joined channels and groups.
    
    Args:
        client: Connected TelegramClient instance
        
    Returns:
        List of channel dictionaries with title and ID
        
    TODO: Use client.get_dialogs() to fetch chats
    TODO: Filter for channels only (exclude private chats)
    TODO: Extract channel title and ID
    TODO: Handle empty dialog list
    TODO: Add channel type classification
    TODO: Add member count information
    """
    # TODO: Call client.get_dialogs()
    # TODO: Filter for Channel type entities
    # TODO: Extract relevant information
    # TODO: Return structured channel list
    return []  # Placeholder return


def select_channel(channels: List[Dict]) -> Dict:
    """
    Interactive CLI prompt to select a channel from the list.
    
    Args:
        channels: List of channel dictionaries from get_user_channels()
        
    Returns:
        Selected channel dictionary
        
    TODO: Display numbered list of channels
    TODO: Handle user input validation
    TODO: Add search/filter functionality
    TODO: Add pagination for large channel lists
    TODO: Add channel preview information
    """
    # TODO: Display channel list with numbers
    # TODO: Get user input
    # TODO: Validate selection
    # TODO: Return selected channel
    return {}  # Placeholder return


def download_messages(client: TelegramClient, channel_id: int, limit: int = 5) -> List[Dict]:
    """
    Download messages from the selected channel.
    
    Args:
        client: Connected TelegramClient instance
        channel_id: ID of the channel to download from
        limit: Maximum number of messages to download
        
    Returns:
        List of message dictionaries
        
    TODO: Use client.iter_messages() to fetch messages
    TODO: Implement pagination for large message counts
    TODO: Extract message content and metadata
    TODO: Handle different message types (text, media, etc.)
    TODO: Add progress reporting
    TODO: Handle rate limiting
    TODO: Add message filtering options
    """
    # TODO: Get channel entity
    # TODO: Iterate through messages
    # TODO: Extract message data
    # TODO: Handle media messages
    # TODO: Add progress tracking
    # TODO: Return structured message list
    return []  # Placeholder return


def save_messages(messages: List[Dict], path: Path) -> None:
    """
    Save downloaded messages to JSON file.
    
    Args:
        messages: List of message dictionaries
        path: Path to save the JSON file
        
    TODO: Convert messages to JSON-serializable format
    TODO: Handle datetime serialization
    TODO: Add metadata to the export
    TODO: Implement pretty printing
    TODO: Add compression options
    TODO: Handle file write errors
    """
    # TODO: Prepare messages for JSON serialization
    # TODO: Add export metadata
    # TODO: Write to file with proper encoding
    # TODO: Handle serialization errors
    pass


def main():
    """
    Main function to run the Telegram client workflow.
    
    TODO: Parse command line arguments
    TODO: Load API credentials
    TODO: Connect to Telegram
    TODO: List and select channel
    TODO: Download messages
    TODO: Save to examples/raw_dump.json
    TODO: Display statistics
    """
    # TODO: Get API credentials from environment or config
    # TODO: Connect to Telegram
    # TODO: Get user channels
    # TODO: Let user select channel
    # TODO: Download messages
    # TODO: Save to file
    # TODO: Display results
    pass


if __name__ == "__main__":
    # TODO: Run main function
    # TODO: Handle exceptions
    # TODO: Add logging
    pass 