requirements.txt
typer>=0.9.0
telethon>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
pathlib2>=2.3.7 